requirements.txt
PyPDF2
customtkinter
tkinterdnd2
PyMuPDF
Pillow
